# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: study
    description: Study coach operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /profile:
    get:
      operationId: getProfile
      tags: [study]
      summary: Get the signed-in student's profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateProfile
      tags: [study]
      summary: Update the signed-in student's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileInput"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /tutor/conversation:
    get:
      operationId: getTutorConversation
      tags: [study]
      summary: Get the student's most recent tutor conversation
      responses:
        "200":
          description: Conversation history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationHistory"
  /dashboard:
    get:
      operationId: getDashboard
      tags: [study]
      summary: Get the student's study dashboard
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dashboard"
  /courses:
    get:
      operationId: listCourses
      tags: [study]
      summary: List study courses
      responses:
        "200":
          description: Courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [study]
      summary: Create a study course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /tutor/messages:
    post:
      operationId: sendTutorMessage
      tags: [study]
      summary: Ask the AI tutor a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TutorMessageInput"
      responses:
        "200":
          description: Tutor response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorMessage"
  /quiz:
    get:
      operationId: getQuiz
      tags: [study]
      summary: Get the current adaptive quiz
      responses:
        "200":
          description: Quiz question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizQuestion"
  /quiz/answers:
    post:
      operationId: submitQuizAnswer
      tags: [study]
      summary: Submit an answer and get feedback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizAnswerInput"
      responses:
        "200":
          description: Answer feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizFeedback"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Profile:
      type: object
      required: [id, fullName, onboardingCompleted]
      properties:
        id: { type: string }
        fullName: { type: ["string", "null"] }
        gradeLevel: { type: ["string", "null"] }
        studyMinutesPerDay: { type: ["number", "null"] }
        learningStyle: { type: ["string", "null"] }
        onboardingCompleted: { type: boolean }
    ProfileInput:
      type: object
      properties:
        fullName: { type: string }
        gradeLevel: { type: string }
        studyMinutesPerDay: { type: number }
        learningStyle: { type: string }
        onboardingCompleted: { type: boolean }
    ConversationMessage:
      type: object
      required: [role, message]
      properties:
        role: { type: string }
        message: { type: string }
    ConversationHistory:
      type: object
      required: [conversationId, messages]
      properties:
        conversationId: { type: ["string", "null"] }
        messages:
          type: array
          items: { $ref: "#/components/schemas/ConversationMessage" }
    CourseTopic:
      type: object
      required: [name, masteryLevel, masteryScore]
      properties:
        name: { type: string }
        masteryLevel: { type: string }
        masteryScore: { type: number }
    Course:
      type: object
      required: [id, name, examDate, level, progress, topics]
      properties:
        id: { type: string }
        name: { type: string }
        examDate: { type: string }
        level: { type: string }
        progress: { type: number }
        topics:
          type: array
          items: { $ref: "#/components/schemas/CourseTopic" }
    CourseInput:
      type: object
      required: [name, examDate, level]
      properties:
        name: { type: string, minLength: 1 }
        examDate: { type: string, minLength: 1 }
        level: { type: string, minLength: 1 }
    StudyTask:
      type: object
      required: [label, duration, kind, completed]
      properties:
        label: { type: string }
        duration: { type: string }
        kind: { type: string }
        completed: { type: boolean }
    Dashboard:
      type: object
      required: [greeting, courseName, courseProgress, strongestTopic, focusTopic, tasks, xp, streak, questionsThisWeek, masteredTopics]
      properties:
        greeting: { type: string }
        courseName: { type: string }
        courseProgress: { type: number }
        strongestTopic: { type: string }
        focusTopic: { type: string }
        tasks:
          type: array
          items: { $ref: "#/components/schemas/StudyTask" }
        xp: { type: number }
        streak: { type: number }
        questionsThisWeek: { type: number }
        masteredTopics: { type: number }
    TutorMessageInput:
      type: object
      required: [message]
      properties:
        message: { type: string, minLength: 1 }
        context: { type: ["string", "null"] }
        conversationId: { type: ["string", "null"] }
    TutorMessage:
      type: object
      required: [role, message, prompt, conversationId]
      properties:
        role: { type: string }
        message: { type: string }
        prompt: { type: string }
        conversationId: { type: string }
    QuizQuestion:
      type: object
      required: [id, number, total, topic, question, options, difficulty]
      properties:
        id: { type: string }
        number: { type: number }
        total: { type: number }
        topic: { type: string }
        question: { type: string }
        options:
          type: array
          items: { type: string }
        difficulty: { type: string }
    QuizAnswerInput:
      type: object
      required: [questionId, answer]
      properties:
        questionId: { type: string }
        answer: { type: string }
    QuizFeedback:
      type: object
      required: [correct, explanation, xp, nextTopic]
      properties:
        correct: { type: boolean }
        explanation: { type: string }
        xp: { type: number }
        nextTopic: { type: string }