# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudyTask } from './studyTask';

export interface Dashboard {
  greeting: string;
  courseName: string;
  courseProgress: number;
  strongestTopic: string;
  focusTopic: string;
  tasks: StudyTask[];
  xp: number;
  streak: number;
  questionsThisWeek: number;
  masteredTopics: number;
}