# lib/api-zod/src/generated/types/profileInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileInput {
  fullName?: string;
  gradeLevel?: string;
  studyMinutesPerDay?: number;
  learningStyle?: string;
  onboardingCompleted?: boolean;
}