# lib/api-zod/src/generated/types/tutorMessage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TutorMessage {
  role: string;
  message: string;
  prompt: string;
  conversationId: string;
}