# lib/api-zod/src/generated/types/tutorMessageInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TutorMessageInput {
  /** @minLength 1 */
  message: string;
  /** @nullable */
  context?: string | null;
  /** @nullable */
  conversationId?: string | null;
}